import: move loading code to class
